+++ src/components/DataTable/DataTable.test.tsx
@@ -1,6 +1,11 @@
+import {
+  SampleData,
+  SampleDataWithError,
+  colDef,
+  generateData,
+} from './test/dataTableTestUtils';
 import { DEFAULT_ELEMENTS_PER_PAGE, DataTable, DataTableProps } from './DataTable';
-import { SampleData, colDef, generateData } from './test/dataTableTestUtils';
-import { render, screen } from 'test/testUtils';
+import { render, screen, waitFor, within } from 'test/testUtils';
 
 const data = generateData();
 const columns = colDef;
@@ -14,6 +19,12 @@ const setup = (props: Partial<DataTableProps<SampleData, unknown>> = {}) => {
   const combinedProps = { ...defaultProps, ...props };
 
   return render(<DataTable {...combinedProps} />);
+};
+
+const waitForRender = async () => {
+  await waitFor(() => {
+    expect(screen.getByRole('table')).toBeInTheDocument();
+  });
 };
 
 // Utility function to check that the whole element is in the table
@@ -32,25 +43,28 @@ const getNumberOfRows = (container: HTMLElement) => {
 };
 
 describe('The DataTable component', () => {
-  it('renders a table', () => {
+  it('renders a table', async () => {
     setup();
-
-    expect(screen.getByRole('table')).toBeInTheDocument();
+    await waitForRender();
   });
 
   describe('the columns', () => {
     const firstElement = data[0];
 
-    it('uses the specified column header', () => {
+    it('uses the specified column header', async () => {
       setup({ data: [firstElement] });
+
+      await waitForRender();
 
       columns.forEach(col => {
         expect(screen.getByText(col.header as string)).toBeInTheDocument();
       });
     });
 
-    it('uses the specified column value', () => {
+    it('uses the specified column value', async () => {
       setup({ data: [firstElement] });
+
+      await waitForRender();
 
       expect(screen.getByRole('table')).toBeInTheDocument();
 
@@ -69,17 +83,23 @@ describe('The DataTable component', () => {
           ],
         });
 
+        await waitForRender();
+
         expect(screen.queryByTestId('sorting_button_name')).not.toBeInTheDocument();
       });
 
-      it('should be present if enabledSorting = true', () => {
+      it('should be present if enabledSorting = true', async () => {
         setup();
+
+        await waitForRender();
 
         expect(screen.getByTestId('sorting_button_name')).toBeInTheDocument();
       });
 
       it('should toggle sort if clicked', async () => {
         const { user } = setup();
+
+        await waitForRender();
 
         // Before clicking it should be not sorted
         expect(screen.getByTestId('head_col_name')).toHaveAttribute(
@@ -114,6 +134,8 @@ describe('The DataTable component', () => {
           ],
         });
 
+        await waitForRender();
+
         // Before clicking it should be not sorted
         expect(screen.getByTestId('head_col_name')).toHaveAttribute(
           'data-sorting',
@@ -123,9 +145,56 @@ describe('The DataTable component', () => {
     });
   });
 
+  describe('when the row has an error message', () => {
+    const elements: SampleDataWithError[] | SampleData[] = generateData(2);
+    elements[0] = {
+      ...elements[0],
+      errorMessages: [{ message: 'ERROR_MESSAGE', status: 'CRITICAL' }],
+    } as SampleDataWithError;
+
+    it('should render the error messages in a row below the main row', async () => {
+      setup({ data: elements });
+
+      await waitForRender();
+
+      checkElementInTable(elements[0]);
+      expect(screen.getByText('ERROR_MESSAGE')).toBeInTheDocument();
+      checkElementInTable(elements[1]);
+    });
+  });
+
+  describe('when the row has an error message with docs link', () => {
+    const elements: SampleDataWithError[] | SampleData[] = generateData(2);
+    elements[0] = {
+      ...elements[0],
+      errorMessages: [
+        {
+          message: 'ERROR_MESSAGE',
+          status: 'CRITICAL',
+          docs_link: 'https://example.com',
+        },
+      ],
+    } as SampleDataWithError;
+
+    it('should render the error messages in a row below the main row', async () => {
+      setup({ data: elements });
+
+      await waitForRender();
+
+      checkElementInTable(elements[0]);
+      expect(screen.getByText('ERROR_MESSAGE')).toBeInTheDocument();
+      const link = screen.getByRole('link');
+      expect(link).toBeInTheDocument();
+      expect(within(link).getByText('Learn more')).toBeInTheDocument();
+      expect(link).toHaveAttribute('href', 'https://example.com');
+    });
+  });
+
   describe('the pagination', () => {
-    it('should display the pagination where current page is the first', () => {
+    it('should display the pagination where current page is the first', async () => {
       setup();
+
+      await waitForRender();
 
       expect(screen.getByTestId('datatable-pagination')).toBeInTheDocument();
       // Prev arrow should be disabled, next arrow should be enabled
@@ -143,38 +212,46 @@ describe('The DataTable component', () => {
       );
     });
 
-    it('should show "DEFAULT_ELEMENTS_PER_PAGE" elements per page by default', () => {
+    it('should show "DEFAULT_ELEMENTS_PER_PAGE" elements per page by default', async () => {
       const { container } = setup();
+
+      await waitForRender();
 
       // There should be only DEFAULT_ELEMENTS_PER_PAGE table rows per page
       expect(getNumberOfRows(container)).toBe(DEFAULT_ELEMENTS_PER_PAGE);
     });
 
-    it('should show the specified elements per page by passing the prop', () => {
+    it('should show the specified elements per page by passing the prop', async () => {
       const { container } = setup({
         elementsPerPage: 10,
       });
+
+      await waitForRender();
 
       // There should be only 10 table rows per page
       expect(getNumberOfRows(container)).toBe(10);
     });
 
     describe('when hidePaginationWhenSinglePage is set', () => {
-      it('does not display pagination if there is only one page', () => {
+      it('does not display pagination if there is only one page', async () => {
         setup({
           data: generateData(1),
           hidePaginationWhenSinglePage: true,
         });
+
+        await waitForRender();
 
         expect(screen.queryByTestId('datatable-pagination')).not.toBeInTheDocument();
       });
     });
 
     describe('when hidePaginationPageSize is set', () => {
-      it('does not display the page size selector', () => {
+      it('does not display the page size selector', async () => {
         setup({
           hidePaginationPageSize: true,
         });
+
+        await waitForRender();
 
         expect(screen.queryByTestId('datatable-pagination')).toBeInTheDocument();
 
@@ -185,10 +262,12 @@ describe('The DataTable component', () => {
     });
 
     describe('when paginationContent is set', () => {
-      it('displays the custom pagination content', () => {
+      it('displays the custom pagination content', async () => {
         setup({
           paginationContent: <div data-testid="custom-pagination-content" />,
         });
+
+        await waitForRender();
 
         expect(screen.getByTestId('custom-pagination-content')).toBeInTheDocument();
       });
@@ -196,45 +275,55 @@ describe('The DataTable component', () => {
   });
 
   describe('when stickyHeader is set', () => {
-    it('applies the CSS classes to make the table header sticky', () => {
+    it('applies the CSS classes to make the table header sticky', async () => {
       setup({
         stickyHeader: true,
       });
+
+      await waitForRender();
 
       expect(screen.getByTestId('head_col_name')).toHaveClass('sticky');
     });
   });
 
   describe('when table is empty', () => {
-    it('should show "No results" label', () => {
+    it('should show "No results" label', async () => {
       setup({
         data: [],
       });
+
+      await waitForRender();
 
       expect(screen.getByText(/No results/)).toBeInTheDocument();
     });
 
-    it('should show custom no result label if passing noResultsLabel prop', () => {
+    it('should show custom no result label if passing noResultsLabel prop', async () => {
       setup({
         data: [],
         noResultsLabel: 'CUSTOM_NO_RESULT',
       });
+
+      await waitForRender();
 
       expect(screen.getByText('CUSTOM_NO_RESULT')).toBeInTheDocument();
     });
   });
 
   describe('the filters', () => {
-    it('should be hidden by default', () => {
+    it('should be hidden by default', async () => {
       setup();
+
+      await waitForRender();
 
       expect(screen.queryByTestId('datatable-filters')).not.toBeInTheDocument();
     });
 
-    it('should render only filters for columns with enableColumnFilter = true', () => {
+    it('should render only filters for columns with enableColumnFilter = true', async () => {
       setup({
         enableFilters: true,
       });
+
+      await waitForRender();
 
       expect(screen.getByTestId('datatable-filter-name')).toBeInTheDocument();
       expect(
@@ -243,10 +332,12 @@ describe('The DataTable component', () => {
     });
 
     describe('when enableFilters = true', () => {
-      it('filters should be present', () => {
+      it('filters should be present', async () => {
         setup({
           enableFilters: true,
         });
+
+        await waitForRender();
 
         expect(screen.getByTestId('datatable-filters')).toBeInTheDocument();
       });
@@ -264,21 +355,25 @@ describe('The DataTable component', () => {
       });
 
       describe('the search box', () => {
-        it('should be hidden by default', () => {
+        it('should be hidden by default', async () => {
           setup({
             enableFilters: true,
           });
+
+          await waitForRender();
 
           expect(
             screen.queryByTestId('datatable-searchbox'),
           ).not.toBeInTheDocument();
         });
 
-        it('should be present if enableSearchBox = true', () => {
+        it('should be present if enableSearchBox = true', async () => {
           setup({
             enableFilters: true,
             enableSearchBox: true,
           });
+
+          await waitForRender();
 
           expect(screen.queryByTestId('datatable-searchbox')).toBeInTheDocument();
         });
@@ -302,15 +397,17 @@ describe('The DataTable component', () => {
   });
 
   describe('when the pagination is disabled', () => {
-    it('does not display the pagination', () => {
+    it('does not display the pagination', async () => {
       setup({ disablePagination: true });
+
+      await waitForRender();
 
       expect(screen.queryByTestId('datatable-pagination')).not.toBeInTheDocument();
     });
   });
 
   describe('when passing in a custom table header', () => {
-    it('renders the custom header, not the standard one', () => {
+    it('renders the custom header, not the standard one', async () => {
       setup({
         customTableHeader: (
           <thead data-testid="custom-table-header">
@@ -322,6 +419,8 @@ describe('The DataTable component', () => {
           </thead>
         ),
       });
+
+      await waitForRender();
 
       expect(screen.getByTestId('custom-table-header')).toBeInTheDocument();
       expect(screen.getAllByRole('columnheader').length).toBe(3);