+++ src/components/TableRowWithNote/index.ts
@@ -1,0 +1,3 @@
+import TableRowWithNote from './TableRowWithNote';
+
+export default TableRowWithNote;